feat: jwt finished

+++ api/src/main/java/com/lugiaweather/api/repository/TelefoneRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.lugiaweather.api.model.Telefone;
 
-public interface TelefoneRepository extends JpaRepository<Telefone, Integer> {
+public interface TelefoneRepository extends JpaRepository<Telefone, Long> {
 
 }
